Keep framing property tests optional
